Churn
Problem Statement:
A bank manager is disturbed about losing credit card customers. The manager requires a solution to keep the customer churn rate low. With the results of this task, the manager can proactively engage customers who are likely to churn so as to retain their business.
Aim
This task attempts to analyze credit card data for a bank and build a model to help the bank manager determine credit card customers who are likely to churn. 

Expected Outcome
Three models would be developed for this classification task and following evaluation, the most suitable model will be recommended for deployment.